refactor: remove unused files and restructure database connection handling
- Deleted .env and init.sql files as part of project cleanup.
- Removed main.go and related handler files to streamline the codebase.
- Updated check_schema.go to directly use environment variables for database connection parameters.
- Eliminated unnecessary modules and files related to database structure fixing.
- Cleaned up configuration and error handling to enhance maintainability.
