--- cmd/fix_db_structure/go.mod
@@ -1,5 +1,0 @@
-module fixdb
-
-go 1.24.1
-
-require github.com/lib/pq v1.10.9 // indirect
