+++ fix_db_structure/go.mod
@@ -1,0 +1,5 @@
+module fixdb
+
+go 1.24.1
+
+require github.com/lib/pq v1.10.9
