+++ fix_db_structure/main.go
@@ -2,21 +2,44 @@ package main
 
 import (
 	"database/sql"
+	"fmt"
 	"log"
+	"os"
 
-	"github.com/go-delve/delve/pkg/config"
 	_ "github.com/lib/pq"
 )
 
 func FixDBStructure() {
-	// Carregar configuração
-	cfg, err := config.LoadConfig()
-	if err != nil {
-		log.Fatalf("Erro ao carregar configuração: %v", err)
+	// Parâmetros de conexão com o banco de dados
+	dbHost := "localhost"
+	dbPort := "5432"
+	dbUser := "postgres"
+	dbPassword := "postgres"
+	dbName := "livros"
+
+	// Permitir sobrescrever com variáveis de ambiente
+	if os.Getenv("DB_HOST") != "" {
+		dbHost = os.Getenv("DB_HOST")
+	}
+	if os.Getenv("DB_PORT") != "" {
+		dbPort = os.Getenv("DB_PORT")
+	}
+	if os.Getenv("DB_USER") != "" {
+		dbUser = os.Getenv("DB_USER")
+	}
+	if os.Getenv("DB_PASSWORD") != "" {
+		dbPassword = os.Getenv("DB_PASSWORD")
+	}
+	if os.Getenv("DB_NAME") != "" {
+		dbName = os.Getenv("DB_NAME")
 	}
 
+	// Criar string DSN
+	dsn := fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=disable",
+		dbHost, dbPort, dbUser, dbPassword, dbName)
+
 	// Conectar ao banco de dados
-	db, err := sql.Open("postgres", cfg.GetDSN())
+	db, err := sql.Open("postgres", dsn)
 	if err != nil {
 		log.Fatalf("Erro ao conectar ao banco de dados: %v", err)
 	}
