+++ internal/domain/models/book.go
